Split the tree-manipulation from initialization, etc. to make for more managable files. Create a NodentCompiler class to reduce references between functions and compilers
