+++ .gitignore
@@ -2,3 +2,4 @@
 /node_modules
 /README.md~
 /wip/
+/deprecated/
