+++ lib/output.js
@@ -27,8 +27,8 @@ function formatParameters(code, node, state, traveler) {
     }
     code.write(null, ') ');
 }
-function formatExpression(code, node, operator, state, traveler) {
-    const needed = PARENTHESIS_NEEDED[node.type];
+function formatExpression(code, node, operator, state, traveler, needed) {
+    needed = needed || PARENTHESIS_NEEDED[node.type];
     if (needed === 0) {
         traveler[node.type](node, state);
         return;
@@ -618,9 +618,13 @@ traveler = {
     BinaryExpression: BinaryExpression = function (node, state) {
         const code = state.code;
         const operator = node.operator;
-        formatExpression(code, node.left, operator, state, this);
+        formatExpression(code, node.left, operator, state, this,
+        		node.left.type==='BinaryExpression' &&
+        		OPERATORS_PRECEDENCE[operator]>=OPERATORS_PRECEDENCE[node.left.operator]);
         code.write(node, ' ', node.operator, ' ');
-        formatExpression(code, node.right, operator, state, this);
+        formatExpression(code, node.right, operator, state, this, 
+        		node.right.type==='BinaryExpression' &&
+        		OPERATORS_PRECEDENCE[operator]>=OPERATORS_PRECEDENCE[node.right.operator]);
     },
     LogicalExpression: BinaryExpression,
     ConditionalExpression: function (node, state) {
