+++ lib/parser.js
@@ -150,12 +150,11 @@ function treeWalker(n,walker,state){
 	return n ;
 }
 
-var comments ;
 function acornParse(code) {
-	comments = [] ;
+	var comments = [] ;
 	var ast = acorn.parse(code,{
 		plugins:{nodent:true},
-		ecmaVersion:6, // TODO: Set from option/config
+		ecmaVersion:6, // TODO: Set from option/config?
 		allowHashBang:true,
 		allowReturnOutsideFunction:true,
 		locations:true,
@@ -175,7 +174,7 @@ function acornParse(code) {
 }
 
 module.exports = {
-		parse: acornParse,
-		treeWalker:treeWalker,
-		_acorn:acorn
+	parse: acornParse,
+	treeWalker:treeWalker,
+	_acorn:acorn
 } ;
